refactor(PUT badges/badge/{badgeID}): all tests written and pass

+++ ferrilata-devilline/Models/DTOs/Input/LevelWithoutHoldersDTO.cs
@@ -12,7 +12,7 @@ namespace ferrilata_devilline.Models.DTOs.Input
         public long LevelId { get; set; }
 
         [JsonProperty(PropertyName = "level")]
-        public long LevelNumber { get; set; }
+        public int LevelNumber { get; set; }
 
         [JsonProperty(PropertyName = "weight")]
         public string Weight { get; set; }
