+++ ferrilata-devilline/Models/DTOs/LevelDTO.cs
@@ -9,7 +9,7 @@ namespace ferrilata_devilline.Models.DTOs
         public long LevelId { get; set; }
 
         [JsonProperty(PropertyName = "level")]
-        public long LevelNumber { get; set; }
+        public int LevelNumber { get; set; }
 
         [JsonProperty(PropertyName = "weight")]
         public string Weight { get; set; }
